feat: se implementado la seccion para crear evento
